refactor: move templates to recipes
